% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phl.R
\name{phl_count}
\alias{phl_count}
\title{Get row count for a Philadelphia table}
\usage{
phl_count(table_name, where = NULL)
}
\arguments{
\item{table_name}{CARTO table name}

\item{where}{Optional WHERE clause}
}
\value{
integer count
}
\description{
Get row count for a Philadelphia table
}
