% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phl.R
\name{phl_crimes}
\alias{phl_crimes}
\title{Get Philadelphia crime incidents}
\usage{
phl_crimes(limit = 1000, where = NULL, year = NULL)
}
\arguments{
\item{limit}{Number of rows (default 1000)}

\item{where}{Optional SQL WHERE clause (without 'WHERE')}

\item{year}{Optional year filter (e.g. 2024)}
}
\value{
tibble: dc_dist, dispatch_date_time, dispatch_date, location_block,
text_general_code, point_x, point_y
}
\description{
Get Philadelphia crime incidents
}
