% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phl.R
\name{phl_permits}
\alias{phl_permits}
\title{Get Philadelphia permits}
\usage{
phl_permits(limit = 1000, where = NULL)
}
\arguments{
\item{limit}{Number of rows (default 1000)}

\item{where}{Optional SQL WHERE clause}
}
\value{
tibble with permit data
}
\description{
Get Philadelphia permits
}
