% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phl.R
\name{phl_query}
\alias{phl_query}
\title{Run a SQL query against Philadelphia open data}
\usage{
phl_query(sql)
}
\arguments{
\item{sql}{SQL query string}
}
\value{
tibble with query results
}
\description{
Uses the CARTO SQL API to query any Philadelphia dataset.
}
