% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phl.R
\name{phl_violations}
\alias{phl_violations}
\title{Get Philadelphia code violations}
\usage{
phl_violations(limit = 1000, where = NULL)
}
\arguments{
\item{limit}{Number of rows (default 1000)}

\item{where}{Optional SQL WHERE clause}
}
\value{
tibble with violation data
}
\description{
Get Philadelphia code violations
}
